all crud apis are successful

+++ controllers/contactController.js
@@ -4,7 +4,7 @@ const Contact =require("../models/contactModel");
 
 //@desc Get all contacts
 //@route Get /api/contact
-//@acess public
+//@acess private
 const getContacts = asyncHandler(async(req,res)=>{
     const contacts= await Contact.find({user_id :req.user.id});
     res.status(200).json(contacts);
@@ -12,7 +12,7 @@ const getContacts = asyncHandler(async(req,res)=>{
 
 //@desc Get particular contact
 //@route Get /api/contact/:id
-//@acess public
+//@acess private
 const getContact = asyncHandler(async (req,res)=>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
@@ -24,7 +24,7 @@ const getContact = asyncHandler(async (req,res)=>{
 
 //@desc Create all contacts
 //@route Post /api/contact
-//@acess public
+//@acess private
 const createContact = asyncHandler(async (req,res)=>{
     console.log("request body is : ", req.body);
     const {name,email,phone} = req.body;
@@ -43,14 +43,18 @@ const createContact = asyncHandler(async (req,res)=>{
 
 //@desc Update Contact
 //@route Put /api/contact/:id
-//@acess public
+//@acess private
 const updateContact = asyncHandler(async (req,res)=>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not found");
     }
-    const {name,email,phone} = req.body;
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User dont have permission to update other user contacts");
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -61,15 +65,22 @@ const updateContact = asyncHandler(async (req,res)=>{
 
 //@desc Delete particular contacts
 //@route Delete /api/contact/:id
-//@acess public
+//@acess private
 const deleteContact = asyncHandler(async (req,res)=>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not found");
     }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User dont have permission to delete other user contacts");
+    }
+
     await Contact.deleteOne();
-    res.status(201).json(contact);
+    res.status(200).json(contact);
+
 })
